pyproject.toml
[project]
name = "geolocation-agent"
version = "0.1.0"
description = "A geolocation agent that determines where a photo was taken through iterative hypothesis testing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langchain-google-genai>=2.0",
    "tavily-python>=0.5",
    "google-search-results>=2.4",
    "googlemaps>=4.10",
    "Pillow>=10.0",
    "exifread>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geolocation_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["live: tests that hit real external APIs (deselect with '-m not live')"]
addopts = "-m 'not live'"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
